refactor #149: move footer to molecules folder
